lib/ukdebug: Add thread info to debug messages
Add the name of the thread to the debug messages.
If the name is null or the thread was not created
with uksched library, the pointer returned from
uk_thread_current is printed.
This will be an invalid pointer if the thread was
not created with uksched library -the boot process
for example.

Signed-off-by: Dafna Hirschfeld <dafna3@gmail.com>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>