lib/ukdebug: Separate condition string from format in UK_ASSERT, UK_WARNIF
Whenever a condition expression for UK_ASSERT() and UK_WARNIF()
is not fulfilled, the macros stringify the condition and print
it with uk_printd(). Because this string was handed over with
the format, the modulo operator (%) in the expression causes
formatting errors.
This patch hands over this condition string as separate argument
which will not be processed by uk_printd().

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Costin Lupu <costin.lupu@cs.pub.ro>