plat/common: Common x86 threading code
We define a new primitive - the software context - used in context switching for
cooperative scheduling and scheduling on PV guests. The code for Xen platform is
reused to enable cooperative scheduling on KVM as well.

Since Unikraft may run different schedulers on dedicated CPUs, a new abstraction
is introduced, ukplat_ctx_callbacks, for enabling support of other types of contexts
as well (e.g. hardware contexts for preemptive scheduling).

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>