lib/uksched: Updates for inlining with plat/common changes
When creating a new thread, the scheduler allocates the stack, sets the thread
address on the stack top and pushes the thread function and argument on the
newly created stack. The context resource will be created by the platform which
will set the stack pointer as provided by the scheduler and the instruction
pointer to a value representing the start routine for the context.

Other changes:
- introducing uk_sched_default_init function
- using new context callbacks abstraction
- new uk_sched_start function replacing uk_sched_run; the start function is the
same for all schedulers
- all schedulers have an idle thread
- revisited semantics: dropping start/stop names for threads functions since we
cannot control exactly when threads start/stop running; more suitable function
names are add/remove

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>