plat/kvm: Place early boot code to the start of the image (x86)
Places early boot code (e.g., 64bit mode switch) and
data (e.g., GDT) to the start of the image.
This avoids general protection faults that were caused
whenever the linker placed such items to addresses
that were not accessible while the mode switch to 64bit
and address space setup was not finished.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Yuri Volchkov  <yuri.volchkov@neclab.eu>