lib/ukdebug: Print bottom of stack instead of thread name
The option of printing the thread name of the one
issueing a uk_printd() call did not work all the
time and causes the system to access invalid memory
addresses for retrieving the thread name string. The
reason is that we currently do not have a fast and
efficient way to differentiate if a stack was created
by libuksched or not (e.g., boot stacks, interrupt stacks).

This patch changes this option to print the bottom
address of the current stack instead. This way, the user
can still distinguish different threads or stack types
by their stack address and we remove any dependency to
other libraries or system configuration.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Yuri Volchkov <yuri.volchkov@neclab.eu>