build: save kconfig dependency files in the separate dir
The *conf utilities are generating a bunch of empty *.h files,
reflecting the configuration menu symbols. The corresponding file is
touched if symbol has been changed. This is used for making smart
decisions which files require recompilation.

So far they have been put in the build/kconfig folder, with other
stuff. Let's save them in the separate dir, to make file structure
a bit more beautiful.

Signed-off-by: Yuri Volchkov <yuri.volchkov@neclab.eu>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>