plat/kvm: Add initial PCI bus driver for x86_64
Introduces the first & initial version of PCI bus driver.
It is able to iterate over PCI devices and detect
if an driver is available. It is going to call a driver
callback function to handover initialization for each
found device.
Although the code is put to `plat/common` it has some
dependency to KVM x86_64 with the I/O functions. These
functions should be moved to the appropriate place on
successive patches for PCI.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Signed-off-by: Sharan Santhanam <sharan.santhanam@neclab.eu>
Reviewed-by: Yuri Volchkov <yuri.volchkov@neclab.eu>