plat/xen: Add new memory region - demand area (x86_64)
Demand area is a new memory region used for Xen for on-demand memory
mappings. An example of use case for on-demand mappins is the grant
subsystem which requires a virtual memory region different than the
memory used for heap.

This patch also introduces the functions for reading and writing page
table entries, needed when creating memory mappings in the demand area.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Yuri Volchkov <yuri.volchkov@neclab.eu>