build: Adjust sed script to avoid treating arm64 as arm
We are using the sed scripts to parse target architecture from
compiler's output. But for arm64, the HOSTARCH and UK_ARCH would
be formatted to arm by mistake. That is because of the order of
sed's scripts:
-e 's/arm64.*/arm64/' -e 's/aarch64.*/arm64/' \
-e 's/arm.*/arm/'
The first line's output "arm64" will be formatted to "arm" by the
second line. So we have to give some reserve characters in second
line to prevent transfer "arm64" to "arm"

Signed-off-by: Wei Chen <Wei.Chen@arm.com>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>