build: Introduce a new variable UK_FAMILY
This patch series modifies the folder layout organized by CPU
families like:
unikraft----arch----arm----arm
              |      |-----arm64
              |
              |-----x86----x86
                     |-----x86_64

In this case, we introduce a new variable UK_FAMILY for build scripts.
we should not expect the user to set or select UK_FAMILY in a menu or
by setting the variable externally. So we retrieve UK_FAMILY directly
from CONFIG_UK_ARCH. Because of this we can keep Compiler.uk, Makefile.uk,
and Config.uk directly under the UK_FAMILY directory.

Signed-off-by: Wei Chen <Wei.Chen@arm.com>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>