arch: Reorganize arch folder to reflect the CPU family schema
As we have introduced UK_FAMILY for build scipts. In order to
reflect the CPU family schema, we modify the "arch/UK_ARCH"
to "arch/UK_FAMILY/UK_ARCH".

The "arch/UK_FAMILY" can be used to store the common codes and build
scripts for the architectures of same family.

For individual architectures, it might make sense to use forward
declarations in Compiler.uk and Makefile.uk to include a corresponding
file in a subfolder. For instance arm/Makefile.uk includes
arm/arm/Makefile.uk when an ARM32 architecture was selected.

Signed-off-by: Wei Chen <Wei.Chen@arm.com>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>