plat/kvm: don't crash on unhandled IRQ
Crashing on unhandled interrupts is not very helpful, because things as
simple as hitting a key on the VGA console leads to a crash (because
there is no keyboard driver yet).

Generally speaking, crashing on an unhandled IRQ is unnecessarily harsh.
If we just ignore it, the worst thing that can happen is that the
hardware that raised the interrupt feels ignored and floods us with
additional interrupts, which could lead to the system stalling; but that
is not any worse than an immediate crash, and in most cases, just
showing an error message and continuing will be completely fine.

Also removed a comment that seems somewhat misleading about what the
function actually does.

Signed-off-by: Florian Schmidt <florian.schmidt@neclab.eu>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>