lib/nolibc: move shared definitions to special include file
libc mandates a bunch of definitions that defined from several header
files. At the same time, these definitions are overlapping, but not
easily solved by including one libc header file from another. As an
example (which is fixed in this patch), including stddef.h from stdio.h
provides stdio.h with the required definitions for size_t, but stddef.h
contains more definitions that stdio.h should *not* provide. If nolibc
provides too much, code that compiles with nolibc (because it expects
these over-generous provision of definitions) will not compile with
other libc implementations any more.

We take the solution found in other libc implementations such as newlib
and musl: use __NEED_x macros and include an internal header file with
definitions that cherry-picks those that we need.

This patch does this for the first three data types. More are to follow.

Signed-off-by: Florian Schmidt <florian.schmidt@neclab.eu>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>