arch: Add arm64 architecture config to menuconfig
Add the Arm64 entry for menuconfig. As different silicon vendors may
have different 64-bit Armv8 SoCs. If we want to add them to Config,
it will be a large list. So we only provide Arm's cortex A53~A75 CPUs
for "Processor Optimization"

If we use MARCH_ as the prefix for Arm64 CPUs as x86, when we select
"generic", the MARCH_GENERIC will conflict with x86's MARCH_GENERIC.
So, we use MARCH_ARM64_ for Arm64 as the prefix.

Current supported Arm64 CPU models:
native, generic, cortex-a53, cortex-a57, cortex-a72, cortex-a73,
cortex-a55 and cortex-a75.

Signed-off-by: Wei Chen <Wei.Chen@arm.com>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>