arch/arm: Add more CPU models to processor optimization list
Current processor optimization list only support generic cortex-a7
and one Allwin cortex-a7 SoC. In this patch, we add all armv7
cortex-a series and two armv8 32-bit CPU to support list.

As arm64 and x86_64, we also add auto-detect host CPU option for arm32,
even it's not likely to be used on arm32 platforms.

New supported armv7 cortex-a CPU models:
armv7-a,native
armv7-a,generic-armv7-a
armv7-a,cortex-a8
armv7-a,cortex-a9
armv7-a,cortex-a5
armv7-a,cortex-a15
armv7-a,cortex-a7
armv7-a,cortex-a12
armv7-a,cortex-a17
armv8-a,cortex-a32(full backward compatibility with Armv7)
armv8-a,cortex-a35(full backward compatibility with Armv7)

The generic-armv7-a would be our default option of armv7 processor.

Signed-off-by: Wei Chen <Wei.Chen@arm.com>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>