lib/nolibc: Fix infinite loop in vsnprintf
Modified the PCHAR macro so that it evaluates its parameters exactly
once on every call. The old macro could lead to infinite loops in cases
such as "while (*p) { PCHAR(*p++); }" when size < 2. This expression was
used when evaluating a %d, and the bug would be immediately triggered
when using vasprintf with a format string containing %d, because
vasprintf calls vsnprintf with length 0 to compute the length of the
resulting string.

Signed-off-by: Razvan Rotaru <razvanrtr@outlook.com>
Reviewed-by: Florian Schmidt <florian.schmidt@neclab.eu>