plat/kvm: Do not clear VGA console when at 0-position
This is a leftover from the original behavior where the screen was
cleared and written from the top again when it reached the bottom, which
was changed to the more common scrolling behavior in patch 1d864df. Not
only is this unnecessary now, but if we ever implement support for a
curses-like system where we can write to arbitrary locations on the
screen, this would produce a bug that clears the screen when it's not
supposed to.

Signed-off-by: Florian Schmidt <florian.schmidt@neclab.eu>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>