plat/kvm: Solve some concurrency issues in _libkvmplat_vga_putc()
As terminal_row and terminal_column are static variables, concurrent or
interleaved execution of _libkvmplat_vga_putc can lead to inconsistent
values, which in the worst case can lead to writing outside the bounds
of the video buffer. By making local copies of those variables, working
on those copies, and copying them back at the end, we reduce the impact.
This can still lead to losing characters or full lines, but we should
never leave the buffer area.

For more robust printing, a solution should target higher up the call
chain, for example, inside coutk, so that buffers printed out cannot
interleave.

Signed-off-by: Florian Schmidt <florian.schmidt@neclab.eu>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>