plat/linuxu: Add linuxu (x86_64) interrupts support
We use signals to emulate the behavior of interrupts on plat/linuxu.

To directly issue syscalls to the Linux kernel, we need to be
ABI-compatible with it. For some type definitions that are not fully
specified by POSIX, but that are used in syscalls (looking at you,
sigset_t and fd_set), we therefore provide our own internal definitions
instead of relying on those provided by a libc implementation that might
be POSIX-compliant, but not ABI-compatible.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Signed-off-by: Florian Schmidt <florian.schmidt@neclab.eu>
Reviewed-by: Sharan Santhanam <sharan.santhanam@neclab.eu>