plat/linuxu: Build with memory.c from plat/common
Patch 300c8309 introduced an API for setting the memory allocator. It
added memory.c to the KVM and Xen platform Makefiles, but not to the
linuxu one (which at that time didn't properly build). Now that the
compile errors are fixed, this missing addition leads to a linking error
when compiling for linux userspace. Adding the corresponding line fixes
that.

Signed-off-by: Florian Schmidt <florian.schmidt@neclab.eu>
Reviewed-by: Sharan Santhanam <sharan.santhanam@neclab.eu>