build: Introduce symbols exporting for each library
In order to avoid symbol conflicts and relaxing on namespacing
of library-internal symbols, each library should provide
an `exportsyms.uk` file in its base directory. This file
lists the name of all symbols that should be exported
by the library, the rest is treated as internal symbol
which cannot be linked by another library. A warning is
thrown by `make` whenever this symbols file is absent.
The scope of each symbol for such a library stays unchanged.
This way, this patch won't break existing Unikraft projects
and libraries.

The location of this file can be overwritten by defining the
path with the [LIBNAME]_EXPORTS variable within the
appropriate `Makefile.uk`.

Signed-off-by: Yuri Volchkov <yuri.volchkov@neclab.eu>
Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Florian Schmidt <florian.schmidt@neclab.eu>