include/uk: import bitcount.h
The definitions in this file are currently present in FreeBSD in the
sys/sys/types.h file with a comment stating that the definitions should
be elsewhere. The same definitions can be found in newlib v2.5.0 in
libc/sys/rtems/include/machine/_bitcount.h which we chose to port in
Unikraft. The original file was slightly adapted to use Unikraft include
guards and types.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Yuri Volchkov <yuri.volchkov@neclab.eu>