include/uk: bitopts.h - remove already existing functions
Some of the functions provided by bitopts.h are already implemented in
unikraft. This patch turns these functions into wrappers around
existing ones.

The original FreeBSD code is using the same (atomic) function for both
__set_bit and set_bit. That is not correct. The underscored version
should be non-atomic.

This patch also drops the test_bit for now, because there is a problem
with its implementation.

Signed-off-by: Yuri Volchkov <yuri.volchkov@neclab.eu>
Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Florian Schmidt <florian.schmidt@neclab.eu>