arch/arm64: Avoid using the floating-point and Advanced SIMD registers
On Arm64, sometimes, the GCC will use floating-point and Advanced
SIMD registers to pass parameters. For example, the va_list will
use the SIMD&FP registers (like q0, q1) to store parameters, no
matter you are using floating-point or not.

But before we include the FP & SIMD registers in context switch,
we should not enable the FP & SIMD access for kernel. So, we use
the GCC -mgeneral-regs-only flag to force GCC to use generic
registers only util the floating-point and Advanced SIMD registers
are required actually.

Signed-off-by: Wei Chen <Wei.Chen@arm.com>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>