include/uk: fix test_bit atomicity
ukarch_test_bit_sync does not make sense. A barrier() call is not a
memory barrier as was said in the comment. Even if it it was, it would
not make ukarch_test_bit_sync atomic.

This commit merges these two function together, and adds a missing
uk_test_bit into bitops.h.

Signed-off-by: Yuri Volchkov <yuri.volchkov@neclab.eu>
Reviewed-by: Florian Schmidt <florian.schmidt@neclab.eu>