build: Support for static libraries
Introduce support for linking (externally built) static libraries.

A static library can be registered at three different places:
 * [LIBNAME]_ALIBS and [LIBNAME]_ALIBS-y
   static libraries used when (pre-)linking a library object
 * [PLATNAME]_ALIBS and [PLATNAME]_ALIBS-y
   static libraries that are linked at the final linking step but only
   for a specific platform
 * UK_ALIBS and UK_ALIBS-y
   static libraries that are linked at the final linking step

While the link order of object files is irrelevant to the linker,
static libraries are handled differently. During linking, static
libraries are only visited once. The only objects taken are those
that resolve at least one yet unresolved symbol. Because of this,
we populate the list of static libraries right after populating all
relevant object files.

Any order or circular dependency within a group of static libraries
is resolved by instructing the linker to scan the archives multiple
times. This is done by the --start-group and --end-group linker flags.

With this change, we also remove the UK_LINK and UK_LINK-y variables
that were used to simplify the platform's linker scripts. Because of
the link order and additional linker flags, we directly assemble the
linker command by taken the items from the corresponding OLIBS and
ALIBS lists.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Felipe Huici <felipe.huici@neclab.eu>