lib/uknetdev: Netdev initialization
Introduce a 'DPDK RTE Ethernet' inspired API for initializing a
Unikraft network device. The initialization is done in the following
order: (1) Configure main aspects of device (e.g., number of
queues), (2) configure each transmit and receive queue, and (3) start
the device. We also introduce an interface for querying the underlying
device capabilities for each configuration step.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Signed-off-by: Razvan Cojocaru <razvan.cojocaru93@gmail.com>
Reviewed-by: Sharan Santhanam <sharan.santhanam@neclab.eu>