plat/drivers: Introduce virtio PCI device
We introduces support for virtio PCI legacy device support in this
patch. The driver register on the PCI bus to probe for device with
QUMRANET vendor id. The virtio PCI bus adds the virtio dev. The
virtio bus is responsible for bridging the virtio device with the
driver. The virtio PCI device provides configuration callback for
the PCI bus operation performed on the virtio device. Currently we
add support to the KVM platform

Signed-off-by: Sharan Santhanam <sharan.santhanam@neclab.eu>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>