lib/vfscore: Add stdin fop
This patch adds the stdin operation inside lib/vfscore/stdio.c.

The stdio is mapped by default to fd 0. This implementation mimics
the normal read behaviour.(read until count bytes or read until VEOF or
\n is met). The '\r' is replaced with '\n' because ukplat_cink returns
'\r' on KVM and Xen.

For every successful call of ukplat_cink the input is printed
to the stdout.

Signed-off-by: Vlad-Andrei Badoiu <vlad_andrei.badoiu@stud.acs.upb.ro>
Reviewed-by: Yuri Volchkov <yuri.volchkov@neclab.eu>