build: resolve common symbols in each lib individually
If a symbol is not initialized, gcc declares it as a weak. It will be
resolved during the linking time.

Common symbols inside a library are fine, but we don't want them to be
resolved together with common symbols from another library.

For example:
lib/foo/main.c:
   int some_variable;
   int main(void) {
      printf("libfoo: %p\n", some_variable);
   }
lib/bar/main.c:
   char some_variable;
   int main(void) {
      printf("libbar: %p\n", some_variable);
   }

After linking the final Unikraft image, both libfoo and libbar will
get one address for their some_variable. Even if neither of them was
never declared "extern". Even if they have different types. Even if
some_variable is not listed in exportsyms.uk

The solution is to link library object files (libname.ld.o) with
'-d'. This will force linker to do allocation for common symbols.

Signed-off-by: Yuri Volchkov <yuri.volchkov@neclab.eu>
Reviewed-by: Sharan Santhanam <sharan.santhanam@neclab.eu>