lib/uklock: Use timeout equal to 0 for trying down a semaphore
Until now, timeout equal to 0 meant waiting until the semaphore counter
would have got greater than zero, potentially forever. We allow now timeout
equal to 0 to be used when the user wants just to try down the semaphore
and return immediately, in order to be compliant with the behavior of the
sem_timedwait() POSIX function.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Yuri Volchkov <yuri.volchkov@neclab.eu>