lib/uksched: Introduce thread attributes
We introduce a thread attributes abstraction which is used for
configuring thread behavior. The set of currently available
attributes was inspired from the pthread interface.

Thread attributes can be set either when creating the threads or
during thread execution by calling the corresponding setters. If
an attribute is not supported by the underlying scheduler then
-EINVAL is returned by setters.

The detach attribute is common to all threads, regardless of the
scheduling policy, and if set it means that the thread resources
will be automatically freed on thread exit. By default, threads
are not detached. The complete detach support will be added in a
following commit.

Time slice and priority make sense only for preemptive schedulers.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Florian Schmidt <florian.schmidt@neclab.eu>