lib/uksched: Add support for waiting threads
Similar to waiting processes, we introduce support for waiting
threads. This is the conventional way of getting notified by thread
termination events and freeing thread resources upon exit.

A thread can be waited on by other threads. If no such behaviour is
desired then one should set the detached attribute for the thread. This
implies that the thread resources will be automatically freed when the
thread exits.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Florian Schmidt <florian.schmidt@neclab.eu>