plat/common: Zero out extregs area when creating new thread
Instructions such as xsave and xsaveopt, which are designed to save the
extended registers on recent x86 CPUs, do not guarantee that they touch
the complete xsave header area, but rather only the bits that they need
to set. This can lead to invalid xsave header areas if the area is not
zeroed out before use, which in turn leads to GP traps on xrstor.
Note that simply zeroing out the area is not enough, because an all-zero
xsave legacy region is also invalid.

Thanks to Radu Nicolau for reporting this bug.

Signed-off-by: Florian Schmidt <florian.schmidt@neclab.eu>
Tested-by: Radu Nicolau <radunicolau102@gmail.com>
Reviewed-by: Yuri Volchkov <yuri.volchkov@neclab.eu>