include/essentials: Fix redefinition warnings when using newlib
For some macrodefinitions, newlib headers do not care if they were
previously defined and this results in redefinitions compiler warnings.
In some way, it makes sense to give priority to newlib's definitions if
one decides to use it as a libc implementation. Therefore we include the
newlib headers in essentials.h to give priority to their definitions
instead of ours.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>