arch/spinlock: Add spin_lock_irq variants
Spin_lock_irq* variants are used for synchronizing data that might be
accessed from within an interrupt handler. Added (void)(lock) to each
of the spin_lock macros to force the compiler to check the existence
of the lock.

The DEFINE_SPINLOCK() macro now defines the variable, and thus it can
be used as `static DEFINE_SPINLOCK(x)`.

Signed-off-by: Cristian Banu <cristb@gmail.com>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>