plat/*: Introduce unikraft internal constructors
This patch adds a new section for the ukplat_ctortab
array. The array is NULL-terminated and consists
of function pointers to constructors. The pointers are
sorted by priority (0-7) and the array is populated at link
time. Libraries can register a constructor function by using
the new macro UKPLAT_CTOR_FUNC() (provided with
include/uk/plat/ctors.h). This patch was needed because
C++ normally uses __attribute__((constructor)) and without
this we would run the C++ constructors during the startup with the
internal unikraft constructors.

This is based on the previous patch of Simon Kuenzer.

Signed-off-by: Vlad-Andrei Badoiu <vlad_andrei.badoiu@stud.acs.upb.ro>
Reviewed-by: Sharan Santhanam <sharan.santhanam@neclab.eu>