lib/nolibc: Provide `ar.h` header
`ar.h` defines macros and data types for working with archive files
created with the Unix `ar` tool. The header is ported from FreeBSD.

The motivation to add it to nolibc is to make it possible to create
simple and flat ar-based filesystems for Unikraft with
nolibc. Furthermore, it does not introduce any costs: Adding `ar.h`
does not introduce any symbols to nolibc. It also seems to be usual
that libc's ship this header as one can see on other libc implementations
(e.g., newlib, musl, glibc).

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Felipe Huici <felipe.huici@neclab.eu>