plat: Fix ostensibly random crashes during boot
By putting the ALIGN inside the .preinit_array and .init_array sections
in the linker scripts, the sections would always be aligned and never
empty, so they would be created even when there were no constructors
inside. This for some reason completely breaks the proper loading of the
data section, leading to a large number of seemingly unrelated bugs
during boot with lots of static data structures with effectively random
content.

I have to admit I don't fully understand all the moving parts here,
especially why this so trashes the contents of the data section. Also,
a minimal example with just a hello world application does not trigger
the problems, but they appear as soon as you replace nolibc with newlib.
It boots again properly if you add a constructor.

Signed-off-by: Florian Schmidt <florian.schmidt@neclab.eu>
Reviewed-by: Felipe Huici <felipe.huici@neclab.eu>