lib/ukmpi: Provide blocking uk_mbox_recv()
Commit a48d634 ("lib/uklock: Use timeout equal to 0 for trying down a
semaphore") implicitly changed the behavior of `uk_mbox_recv_to()`. When
passing timeout `0`, the function is not blocking until a message
arrives but trying to receive a message and returning immediately.

This patch introduces `uk_mbox_recv()` which brings back this blocking
behavior. I chose introducing this new function in order to keep the
mailbox API inline with the semaphore ones.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Costin Lupu <costin.lupu@cs.pub.ro>