plat/kvm: Add initrd fields to configuration struct
In order to remember where an initrd was found in memory, we add
region definitions to the configuration struct. One definition is used
to remember where the initrd starts and ends, another one is used to
remember the second part of the heap (`heap2`).
Unfortunately, at least on x86, the initrd may be placed somewhere in
the middle of the memory range that is currently detected as heap. In
such a case the heap has to be split into two parts: the area before
the initrd and the area after. This avoids copying or mapping the
initrd to a different location during boot.
The idea is that heap2 is only filled out when a split happened
and two heap parts are left. Otherwise, this field is not used.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Felipe Huici <felipe.huici@neclab.eu>