plat/kvm: x86: Retrieve initrd location from Multiboot
Inspects Multiboot modules information to retrieve the location of an
loaded initramdisk. It will most likely be placed by Multiboot within the
region that we announce as usable for heap. In such a case, we cut out
the initrd from the heap range and report two heap regions: one before the
initrd, one after the initrd.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Felipe Huici <felipe.huici@neclab.eu>