build/rules: Add support for pre-processing rules
Extends the build system with pre-processing rules. Any source file
can be pre-processed before it gets compiled. For this purpose, the
syntax is extended with an optional '>' operator followed with the
target file extension after pre-processing. As an example,

LIB_SRCS-y += preprocess.m4>.c

means that the source file `preprocess.m4` will be pre-processed to
`preprocess.c` (placed in the build directory) before getting
compiled. The according compiler rule is selected depending on the
given target extension. In the example the C compiler is invoked
afterwards to produce `preprocess.o`.

Similar to the compile rule selection, a switch statement is prepared
where pre-process rules will be selected depending on the original
source file extension.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Felipe Huici <felipe.huici@neclab.eu>