build: add support to specify linking-stage makefile
By setting UK_PLAT_*_LINKER, a platform library can now set the path and
filename of the Linker.uk file. The default is the old behavior of
"Linker.uk" in the platform library base directory. This can be necessary
if a platform library effectively builds binaries for multiple
closely-related platforms: for example, libkvmplat creating a binary
with a multiboot header for QEMU, and one without (and a different entry
point) for Firecracker.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Signed-off-by: Florian Schmidt <florian.schmidt@neclab.eu>
Reviewed-by: Felipe Huici <felipe.huici@neclab.eu>