plat/xen: Align x86 interrupt stack address at compile time
This patch aligns the stack base address at compile time and makes the
irqstack variable public. Before this patch, the IRQ stack size was
2 * STACK_SIZE bytes because the aligned address was computed at runtime
and the bottom half of the stack was wasted. This patch fixes that and
aligns the stack address at build time, therefore avoiding a waste of
STACK_SIZE bytes.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Florian Schmidt <florian.schmidt@neclab.eu>