lib/ukschedcoop: Fix yield behavior when current thread is first in thread list
When the current thread is the only one in the thread list, if it wants
to yield after it creates a new thread then schedcoop_schedule() will
choose the same current thread (instead of the new one) because it is
the first one in the running list. This issue violates the expected
behavior of Round-Robin cooperative schedulers. The solution is to avoid
adding the current thread to the thread list as long as it is running. This
way the first thread in the running list is always the thread which will be
scheduled next.

However, this solution is not complete. If we have a single thread
(hence, the threads list is empty) and it wants to sleep for some time,
schedcoop_schedule() will enter in an infinite loop trying to find the
next thread. Therefore, the solution addition is to keep the sleeping
threads on a separate list: in scenarios like this when the "ready"
threads list is empty, schedcoop_schedule() will iterate over the
sleeping threads until one of them will wake up.

The described problem should also occur on Mini-OS, where this
implementation was ported from.

This patch also adds scheduler callbacks to be called when threads block
or wake up. On blocking, the cooperative scheduler will add the threads
in the sleeping threads list whenever a thread will want to sleep
(wakeup_time > 0). On waking up, the sleeping threads will be removed
from the sleeping threads list and added to the "ready" threads list.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Florian Schmidt <florian.schmidt@neclab.eu>