lib/ukschedcoop: Fix race condition when trying to wake up current thread
The previous patch introduced a race condition: if the current thread
blocks and yields, and an interrupt is triggered whose handle tries to
wake up the current thread then the thread will remain blocked until the
next interrupt (potentially forever) because the current thread cannot
be added to the ready threads list.

The solution is to set a flag on the current thread as soon as the next
thread is scheduled. The interrupt will queue the current thread if and
only if this flag is set.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Florian Schmidt <florian.schmidt@neclab.eu>