plat/kvm: 4k-granularity page table for the first 2MB
The main effect is that trying to read or write from the null pointer or
very low addresses (that are almost invariably errors) now leads to a
page fault instead of silently writing to addresses. Basically, the
behavior C programmers expect when writing userspace applications. The
only holes we punch in the first MB are for where multiboot puts its
information, and for the VGA buffer.

Signed-off-by: Florian Schmidt <florian.schmidt@neclab.eu>
Signed-off-by: Sharan Santhanam <sharan.santhanam@neclab.eu>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>