build: always produce 2 images: w/ and w/o debug syms
On contrary to previously existing scheme, the *.dbg image is a fully
functional binary now. A user still can run qemu/xen with a stripped
image, and attach gdb to it, specifying the *.dbg image as a source.

A full dbg images are not making more overhead comparing to
'--only-keep-debug' ones (apart from disk space), but are bringing
convenience.

Another effect of this patch - *.dbg file is produced
always (menuconfig is gone). The 'strip' operation does not cost us
much, and the stripped image equivalents to image build without debug
symbols. But even in production systems it is handy to have an image
with debug symbols by your side to perform a post-mortem analysis of a
crash.

This approach allows to have strip-able sections, existing only in the
*.dbg files. Now user can append the variable `STRIP_SECTIONS_FLAGS-y`
in the library/application makefile, to list all the section he does
not need in the final image.

Signed-off-by: Yuri Volchkov <yuri.volchkov@neclab.eu>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>