plat: prepare linker script for thread-local storage
Aggregate all .tdata and .tbss variables in the correct sections, and
define global variables that point at section beginning and end.

Note that because the .tbss section is zero-size in the ELF file, just
setting _tls_end to the end of it doesn't give us the the size of the
memory area once loaded, so we use SIZEOF to have it point to the end.

Signed-off-by: Florian Schmidt <florian.schmidt@neclab.eu>
Reviewed-by: Costin Lupu <costin.lupu@cs.pub.ro>